feat: :sparkles: Redesigned the manga details page and improved safety checks due to errors found in testing.
The manga details page now makes use of a navbar at the bottom of the page to traverse between viewable options such as chapters, details and recommendations. Improved safety checks on the homepage to not render some section if the data is not there. Also changed the skeletons a bit.